third_party/blink/web_tests/fast/canvas/OffscreenCanvas-commit-frameless-doc-expected.txt

This is a testharness.js-based test.
[FAIL] Verify that the placeholder canvas can be used as an image source
  assert_array_equals: expected property 2 to be 255 but got 0 (expected array [0, 0, 255, 255] got object "0,0,0,0")
Harness: the test ran to completion.
